+++ incapsula/client_mtls_client_to_imperva_certificate_test.go
@@ -62,6 +62,9 @@ func TestGetClientCaCertificateBadJSON(t *testing.T) {
 }
 
 func TestGetClientCaCertificateInvalidApiConfig(t *testing.T) {
+	restore := withShortRetries()
+	defer restore()
+
 	apiID := "foo"
 	apiKey := "bar"
 	certificateID := "42"